+++ src/plugins/BuiltinPlugin.js
@@ -227,7 +227,19 @@ export const regexRules = [
         ],
         link: "https://github.com/ilib-js/ilib-lint/blob/main/docs/source-no-manual-date-formatting.md",
         severity: "error"
-    }
+    },
+    {
+        type: "resource-matcher",
+        name: "resource-snake-case",
+        description: "Ensure that when source strings contain only snake case (words and/or numbers separeated by underscores) and no whitespace, then the targets are the same",
+        note: "Do not translate the source string if it consists solely of snake-cased strings and/or digits. Please update the target string so it matches the source string.",
+        regexps: [
+            "^\\s*[a-zA-Z0-9]*(_[a-zA-Z0-9]+)+\\s*$",
+            "^\\s*[a-zA-Z0-9]+(_[a-zA-Z0-9]+)*_\\s*$"
+        ],
+        link: "https://gihub.com/ilib-js/ilib-lint/blob/main/docs/resource-snake-case.md",
+        severity: "error"
+    },
 ];
 
 // built-in ruleset that contains all the built-in rules
@@ -254,7 +266,8 @@ export const builtInRulesets = {
         "resource-no-space-between-double-and-single-byte-character": true,
         "resource-no-halfwidth-kana-characters": true,
         "resource-no-double-byte-space": true,
-        "resource-no-space-with-fullwidth-punctuation": true
+        "resource-no-space-with-fullwidth-punctuation": true,
+        "resource-snake-case": true,
     },
 
     source: {